AABB, DistanceProxy, Simplex, TreeNode: Made more methods Aggressively Inlined to try to help DynamicTree speed Body: Fix from Box2D where Contacts were reporting Touching == false when a body is teleported onto another body. Also Added Transform and Position Properties to make it more .net-ish ContactListener: Added XMLDoc ContactManager & Contact: Used Properties instead of Get methods ContactSolver, PrismaticJoint: Removed redundant (float) casts DynamicTree: Moved some duplicate code, added "in" hints on method parameters, added AggressiveInlining Island: Removed IDisposable because we're using C#, and we never use Island in an IDisposable manner. RevoluteJoint: Removed redundant "bool active = false;" World: Fixed XMLDoc

+++ src/box2dx/Box2D.NetStandard/Dynamics/Contacts/ContactSolver.cs
@@ -409,8 +409,8 @@ namespace Box2D.NetStandard.Dynamics.Contacts
           float vn1 = Vector2.Dot(dv1, normal);
           float vn2 = Vector2.Dot(dv2, normal);
 
-          Vector2 b = new Vector2((float) (vn1 - cp1.velocityBias),
-                                  (float) (vn2 - cp2.velocityBias));
+          Vector2 b = new Vector2(vn1 - cp1.velocityBias,
+                                  vn2 - cp2.velocityBias);
 
           // Compute b'
           b -= Vector2.Transform(a, vc.K); // Common.Math.Mul(vc.K, a);
